Move acl insert from select helper
This move was made to avoid circular imports in order to access the same
functions from different modules.
